+++ src/mruby_integration/models/cursor.cpp
@@ -7,6 +7,8 @@
 
 struct RClass* Cursor_class;
 
+bool cursor_currently_enabled = true;
+
 auto
 mrb_cursor_show(mrb_state*, mrb_value) -> mrb_value
 {
@@ -27,6 +29,28 @@ mrb_cursor_hidden(mrb_state*, mrb_value) -> mrb_value
   return mrb_bool_value(IsCursorHidden());
 }
 
+auto
+mrb_cursor_enable(mrb_state*, mrb_value) -> mrb_value
+{
+  EnableCursor();
+  cursor_currently_enabled = true;
+  return mrb_nil_value();
+}
+
+auto
+mrb_cursor_disable(mrb_state*, mrb_value) -> mrb_value
+{
+  DisableCursor();
+  cursor_currently_enabled = false;
+  return mrb_nil_value();
+}
+
+auto
+mrb_cursor_disabled(mrb_state*, mrb_value) -> mrb_value
+{
+  return mrb_bool_value(!cursor_currently_enabled);
+}
+
 void
 append_models_Cursor(mrb_state* mrb)
 {
@@ -39,4 +63,10 @@ append_models_Cursor(mrb_state* mrb)
     mrb, Cursor_class, "hide", mrb_cursor_hide, MRB_ARGS_NONE());
   mrb_define_class_method(
     mrb, Cursor_class, "hidden?", mrb_cursor_hidden, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Cursor_class, "enable", mrb_cursor_enable, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Cursor_class, "disable", mrb_cursor_disable, MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Cursor_class, "disabled?", mrb_cursor_disabled, MRB_ARGS_NONE());
 }